fix: add constant file and removed unused imports
