+++ packages/subgraph/src/util.ts
@@ -1,28 +1,19 @@
-import { Address, BigInt, dataSource } from "@graphprotocol/graph-ts";
+import { Address } from "@graphprotocol/graph-ts";
 import { Account } from "../generated/schema";
+import {
+  MAINNET_SHORT_HELPER_ADDR,
+  ROPSTEN_SHORT_HELPER_ADDR,
+  LOCALHOST_SHORT_HELPER_ADDR,
+  RA_SHORT_HELPER_ADDR,
+  BIGINT_ZERO,
+} from "./constants";
 
-export const BIGINT_ONE = BigInt.fromI32(1);
-export const BIGINT_ZERO = BigInt.fromI32(0);
-export const SHORT_HELPER_ADDR = getShortHelperAddr(dataSource.network());
-const MAINNET_SHORT_HELPER_ADDR = Address.fromString(
-  "0x3b4095D5ff0e629972CAAa50bd3004B09a1632C5"
-);
-const ROPSTEN_SHORT_HELPER_ADDR = Address.fromString(
-  "0x8903918DFE74476E90B63061E5b9c3E63b65d3F4"
-);
-const GOCHAIN_SHORT_HELPER_ADDR = Address.fromString(
-  "0x9A9f2CCfdE556A7E9Ff0848998Aa4a0CFD8863AE"
-);
-const RA_SHORT_HELPER_ADDR = Address.fromString(
-  "0x5A30a1E3873A2B5Fc9DB9b2b52491C4b6086FAe0"
-);
-
-function getShortHelperAddr(networkName: string): Address {
+export function getShortHelperAddr(networkName: string): Address {
   let addr = MAINNET_SHORT_HELPER_ADDR;
   if (networkName == "ropsten") {
     addr = ROPSTEN_SHORT_HELPER_ADDR;
   } else if (networkName == "localhost") {
-    addr = GOCHAIN_SHORT_HELPER_ADDR;
+    addr = LOCALHOST_SHORT_HELPER_ADDR;
   } else if (networkName == "rinkebyArbitrum") {
     addr = RA_SHORT_HELPER_ADDR;
   }
